upadated Gemma3InterleaveEmbeddings

+++ keras_hub/src/models/gemma3/gemma3_interleave_embeddings.py
@@ -81,7 +81,7 @@ class Gemma3InterleaveEmbeddings(keras.layers.Layer):
         # later.
         zeroth_index_text_embeddings = ops.take(
             flat_text_embeddings,
-            indices=ops.squeeze(to_add, axis=-1),
+            indices=ops.cast(ops.squeeze(to_add, axis=-1), "int32"),
             axis=0,
         )
